+++ com.unity.render-pipelines.high-definition/Runtime/RenderPipeline/XR/XRSystem.cs
@@ -137,6 +137,12 @@ namespace UnityEngine.Rendering.HighDefinition
                     // Disable vsync on the main display when rendering to a XR device
                     QualitySettings.vSyncCount = 0;
 
+                    if(display != null)
+                    {
+                        display.zNear = camera.nearClipPlane;
+                        display.zFar = camera.farClipPlane;
+                    }
+
                     CreateLayoutFromXrSdk(camera, singlePassAllowed);
                 }
                 else